// template/react-starter-kit/src/.generated/api/model/logItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { LogErrorInfoDto } from './logErrorInfoDto';
import type { LogItemDtoRequestBody } from './logItemDtoRequestBody';
import type { LogItemDtoResponseBody } from './logItemDtoResponseBody';

export interface LogItemDto {
  id: string;
  requestId: string;
  createdAt: string;
  method: string;
  url: string;
  statusCode: number;
  duration: number;
  /** @nullable */
  ip: string | null;
  /** @nullable */
  userAgent: string | null;
  level: string;
  /** @nullable */
  emailHash: string | null;
  /** @nullable */
  requestBody: LogItemDtoRequestBody;
  /** @nullable */
  responseBody: LogItemDtoResponseBody;
  /** @nullable */
  errorInfo: LogErrorInfoDto | null;
}